Adding conditional install

+++ downloadFiles.sh
@@ -5,26 +5,30 @@
 
 if [ $SHELL != '/usr/bin/zsh' ]; then
     sudo usermod -s "$(type -P zsh)" "$(whoami)"
-    echo "✔ zsh as default shell for leslie"  
+    echo "✔ [zsh] as default shell for leslie"  
 fi
 # install tmux
 [ -d "${HOME}/.tmux/plugins/tpm" ] || git clone https://github.com/tmux-plugins/tpm "${HOME}/.tmux/plugins/tpm"
-echo "✔ tmux configured!"
+echo "✔ [tmux] configured!"
 # install gnome-terminal themes
 export TERMINAL=gnome-terminal
-terminal_install_dir="/tmp/gogh"
-info "[gnome-terminal] Configure"
-rm -rf "$terminal_install_dir" && mkdir -p "$terminal_install_dir"
-git clone https://github.com/Mayccoll/Gogh.git "$terminal_install_dir"
-info "[gnome-terminal][configure] One Dark"
-"$terminal_install_dir/themes/one-dark.sh"
-info "[gnome-terminal][configure] Nord"
-"$terminal_install_dir/themes/nord.sh"
+terminal_install_dir="${HOME}/gogh"
+if [ -d "${terminal_install_dir}" ]; then
+    echo "✔ [gnome-terminal] already configured!"
+else
+    echo "[gnome-terminal] Configure"
+    rm -rf "/tmp/gogh"
+    rm -rf "$terminal_install_dir"  && mkdir -p "$terminal_install_dir"
+    git clone https://github.com/Mayccoll/Gogh.git "$terminal_install_dir"
+    echo "✔ [gnome-terminal] configured One Dark"
+    "$terminal_install_dir/themes/one-dark.sh"
+    echo "✔ [gnome-terminal] configured Nord"
+    "$terminal_install_dir/themes/nord.sh"
+fi
 # install oh-my-zsh
 ZSH="${HOME}/.oh-my-zsh"
 ZSH_CUSTOM="${ZSH}/custom"
 [ -d "${ZSH}" ] || git clone https://github.com/robbyrussell/oh-my-zsh.git "${ZSH}"
-info "[ohmyzsh][configure] Download plugins"
 declare -A plugins=(
     ["plugins/fast-syntax-highlighting"]="https://github.com/zdharma/fast-syntax-highlighting"
     ["plugins/zsh-syntax-highlighting"]="https://github.com/zsh-users/zsh-syntax-highlighting.git"
@@ -39,9 +43,13 @@ for path in "${!plugins[@]}"; do
     fi
 done
 
-echo "✔ oh my zsh configured!"
+echo "✔ [oh-my-zsh] configured!"
 
 # Install keepass 2
-sudo apt-add-repository ppa:jtaylor/keepass -y
-sudo apt-get install keepass2 -y
-echo "✔ Keepass installed!"
+if type -p keepass2 > /dev/null; then
+    echo "✔ [Keepass] already installed!"
+else
+    sudo apt-add-repository ppa:jtaylor/keepass -y
+    sudo apt-get install keepass2 -y
+    echo "✔ [Keepass] installed!"
+fi
